fix consensus failure
When we migrated to cometBFT 0.37, the new API requires to do the
block execution process in two steps: 1) execute 2) save

Our statedb was not designed to work in two steps, but one.
So there was a bug on the root hash returned by the execute step,
which was different from the one returned by save. That created
a consensus failure if the node was stop when processing transactions.
