+++ vochain/apptest.go
@@ -105,7 +105,11 @@ func (app *BaseApplication) AdvanceTestBlock() {
 	// finalize block
 	app.endBlock(time.Now(), height)
 	// save the state
-	_, err := app.CommitState()
+	_, err := app.State.PrepareCommit()
+	if err != nil {
+		panic(err)
+	}
+	_, err = app.CommitState()
 	if err != nil {
 		panic(err)
 	}
