+++ vochain/start.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"go.vocdoni.io/dvote/config"
+	"go.vocdoni.io/dvote/crypto/ethereum"
 	vocdoniGenesis "go.vocdoni.io/dvote/vochain/genesis"
 
 	tmcfg "github.com/cometbft/cometbft/config"
@@ -204,6 +205,17 @@ func newTendermint(app *BaseApplication,
 		return nil, fmt.Errorf("cannot create validator key and state: (%v)", err)
 	}
 	pv.Save()
+	if pv.Key.PrivKey.Bytes() != nil {
+		fmt.Printf("privkey: %s\n", hex.EncodeToString(pv.Key.PrivKey.Bytes()))
+		signer := ethereum.SignKeys{}
+		if err := signer.AddHexKey(hex.EncodeToString(pv.Key.PrivKey.Bytes())); err != nil {
+			return nil, fmt.Errorf("cannot add private validator key: %w", err)
+		}
+		app.NodeAddress, err = NodePvKeyToAddress(pv.Key.PubKey)
+		if err != nil {
+			return nil, fmt.Errorf("cannot create node address from pubkey: %w", err)
+		}
+	}
 
 	// nodekey is used for the p2p transport layer
 	nodeKey := new(p2p.NodeKey)
